Added simple error handling when the REST call fails.
In case the REST call fails when the main or the details activity is active, a
message will be logged and also a toast will be shown to the user.

In case none of this two activities are active, only the log will be added.

Very simple error handling, it is mainly added to avoid the situation that the
client fails when it is not possible to start a VM.