There was a chance the events handler will load nothing
The problem was tha the "lastEventId" could get out of sync with whatever was in DB.
Since it is a cheep operation to get this one int from DB loading it all the
time on each poll (once in couple of minutes at worst). This way it will never get out of sync.