Explicitly add "Version" header to all API calls
Version 4 of oVirt engine will support versions 3 and 4 of the API.
Version 3 will be identical to the one provided by version 3.6 of the
engine, but version 4 will be different and not backwards compatible.
Users of the API can select what version to use with the "Version" HTTP
header or with the "/v3" or "/v4" URL suffixes. If the version isn't
selected explicitly then the server will automatically select a version,
by default version 4. This patch changes moVirt to explicitly request
version 3 of the API, using the "Version" header.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>